promote SkewProduct class to base
